feat(aws): add reference to install Open Search Service
Update the AWS blueprint to include the option to install the Open Search
Service (Elasticsearch compatible) ACK operator. Also include the required
changes in network security policies to CPS1 Workspaces connect to a
private managed Open Search.

Enabled with the variable "ack_enable_oss". Also updates the outputs to
show the private subnet IDs, as it is required to provision an Open Search
Service domain.